Summarize Tract Genotypic Frequencies by delNero2022 Catchment

This function calculates the genotypic frequencies for a given query allele within census tracts based on the delNero2022 catchment areas. It performs the following steps:

Usage

summarize_tract_genotypic_frequencies_by_delNero2022_catchment(query_allele)

Arguments

query_allele

The allele for which genotypic frequencies are calculated.

Value

A list containing the summarized data, nmdp frequencies, and tract data.

Details

Imports nmdp data for the query allele.

Checks if there are values in the nmdp data.

Calculates nmdp frequencies for each census tract for all states.

Imports delNero2022 catchment areas and formats them appropriately.

Pulls in state geographic area including tract data with centroids.

Joins the delNero2022 catchment areas with the tract centroids.

Joins the nmdp info by tract to the tract geographic info.

Summarizes the nmdp data per catchment area based on tract data.

Returns a list of multiple objects including the summarized data.