NMDP HLA Frequencies Adjusted by US 2020 Census

A dataset containing NMDP allele frequencies adjusted by race according to the 2020 US Census. The data provides a US-level summary of race-adjusted allelic frequencies, including calculated genotypic frequencies.

nmdp allele frequencies adjusted by race according to the US census. This table contains those race-adjusted alleleic frequencies summed up so we have a us-level summarization

Usage

nmdp_hla_frequencies_us_2020_census_adjusted

nmdp_hla_frequencies_us_2020_census_adjusted

Format

An object of class tbl_df (inherits from tbl, data.frame) with 2341 rows and 6 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 2341 rows and 6 columns.

Source

See data-raw/adjust_nmdp_for_us_census.R

data-raw/adjust_nmdp_for_us_census.R

Fields

region

The region considered (e.g., US-level).

loci

The HLA loci (e.g., A, B, or C).

allele

The specific HLA allele.

af

US Census-adjusted allelic frequency, calculated as the sum of race-adjusted frequencies.

calc_gf

Genotypic frequencies calculated from the adjusted allelic frequencies.

hla_source

The source of the data, indicating it was derived from us_census_adjusted_nmdp_summed_by_race.

region

What region was considered

loci

-what loci A,B, or C

allele

What allele

af

US-census adjusted alleleic frequency (sum of us-census adjsuted by race afs)

calc_gf

genotypic frequencies calcualted from the af

hla_source

us_census_adjusted_nmdp_summed_by_race